+++ packages/devtools/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @ai-sdk/devtools
 
+## 0.0.1-beta.6
+
+### Patch Changes
+
+- Updated dependencies [2625a04]
+  - @ai-sdk/provider@3.0.0-beta.31
+
+## 0.0.1-beta.5
+
+### Patch Changes
+
+- Updated dependencies [cbf52cd]
+  - @ai-sdk/provider@3.0.0-beta.30
+
+## 0.0.1-beta.4
+
+### Patch Changes
+
+- Updated dependencies [9549c9e]
+  - @ai-sdk/provider@3.0.0-beta.29
+
+## 0.0.1-beta.3
+
+### Patch Changes
+
+- Updated dependencies [9061dc0]
+  - @ai-sdk/provider@3.0.0-beta.28
+
+## 0.0.1-beta.2
+
+### Patch Changes
+
+- Updated dependencies [366f50b]
+  - @ai-sdk/provider@3.0.0-beta.27
+
 ## 0.0.1-beta.1
 
 ### Patch Changes